refactor(app): move ProtocolSetupStep to organisms
Now nothing has to import it out of pages!
