+++ app/src/organisms/ODD/ProtocolSetup/index.ts
@@ -6,5 +6,6 @@ export * from './ProtocolSetupModulesAndDeck'
 export * from './ProtocolSetupOffsets'
 export * from './ProtocolSetupParameters'
 export * from './ProtocolSetupSkeleton'
+export * from './ProtocolSetupStep'
 
 export type * from './types'
